+++ src/Interfaces/IHatsIdUtilities.sol
@@ -19,13 +19,13 @@ pragma solidity >=0.8.13;
 interface IHatsIdUtilities {
     function buildHatId(uint256 _admin, uint16 _newHat) external pure returns (uint256 id);
 
-    function getHatLevel(uint256 _hatId) external view returns (uint8);
+    function getHatLevel(uint256 _hatId) external view returns (uint32);
 
     function isTopHat(uint256 _hatId) external view returns (bool);
 
-    function getAdminAtLevel(uint256 _hatId, uint8 _level) external view returns (uint256);
+    function getAdminAtLevel(uint256 _hatId, uint32 _level) external view returns (uint256);
 
-    function getTreeAdminAtLevel(uint256 _hatId, uint8 _level) external pure returns (uint256);
+    function getTreeAdminAtLevel(uint256 _hatId, uint32 _level) external pure returns (uint256);
 
     function getTophatDomain(uint256 _hatId) external view returns (uint32);
